Completely rework the CSV methodology:
* Add format to query cli args, validate parser exists for format
* Create a new package under csv/parsers
* Create a generalized Parser interface with the necessary function requirements
* Create a generalized CsvRow interface with a function to return an array of strings
* Move the ParsingGroup interface into the parsing package, generalize it a bit more
* Move all accointing specific code underneath the csv/parsers/accointing package
* Create an AccointingParser interface that defines Parser methods
* Rework logic of top level CSV ParseForAddress:
  * Get parser at runtime by key, initialize its parsing groups
  * Get DB txes and events
  * Pass them to the parser for processing
  * Return []CsvRow as received by Parser
* Rework CSV write method: Call GetRowForCsv on CsvRow to get string array
