+++ csv/parsers/accointing/helpers.go
@@ -1,4 +1,4 @@
-package csv
+package accointing
 
 import (
 	"fmt"
